Add support for area-node:{node}
Areas are taken from labels that start with "area-" (casing doesn't matter).
Inside an area path, anything that isn't a a digit or letter is considered a
separator. So a label `area-CodeGen-Mono` is interpreted as an area path of
`CodeGen-Mono` which has two nodes `CodeGen` and `Mono`.

We now have three ways to match areas:

    area:{area-path}
    area-under:{area-path}
    area-node:{area-node}

The first two take an area path, such as `System.Runtime.InteropServices`.
Let's say you have 1 issue that is labelled with `area-System.Runtime` and two
other issues that are labelled with `area-System.Runtime.InteropServices`.

Here is what these queries would return:

    Query                                     | #Results
    ------------------------------------------|---------
    area:System                               | 0
    area:System.Runtime                       | 1
    area:System.Runtime.InteropServices       | 2
    area-under:System                         | 3
    area-under:System.Runtime                 | 3
    area-under:System.Runtime.InteropServices | 2
    area-node:System                          | 3
    area-node:Runtime                         | 3
    area-node:InteropServices                 | 2

Closes #1.