Make keywords data driven
In order to avoid having to duplicate the keyword information in three places
(the query parsing, the completion, and the syntax highlighter) we need to make
the query parser data driven so that we can list the keywords. Sadly this makes
the query parser a bit more convoluted, but it will make things a lot easier to
maintainer.