WIP: Add capability to edit the current query on the Blazor side
This also fixes navigating to home on the Index page, which resets the query
to the default query. Right now, the URL changes but neither the results nor
the code editor reflect that.

However, it's not that simple. This has introduced another problem where fast
editing of the code editor now cause client/server back and forth, probably
due to out-of-sync network requests between them. This can result in
flickering text, alternating queries, or the inability to edit the query.

You can repro this a 100% by banging the keyboard hard & fast. This sucks, I
don't know how to fix this issue at the moment.