Fix logic of --starting-repo
Right now, specifying --starting-repo has no effect because we need to specify both
--reindex and --starting-repo. The logic is written such that --reindex will
cause all repos to be reindexed, no matter whether we reached the starting
repo or not.

This fixes it by initializing reachedStartingRepo to false unless --reindex is
specified and --starting-repo is not. This means inside the loop we can just
check reachedStartingRepo to see whether the repo is supposed to be
incrementally updated.