Do not refresh search results when index changes
For now, we're no longer refreshing search results when the index changes. The
reason being that the UI has a lot more state now, such as expanded groups
when the query is grouped. In order to have a good UX we would have keep the
state around when refreshing in order to avoid messing with the user's UI.

Refreshing the page would be useful for dashboards that are permanently open
in the browser but I'm not aware of anyone doing this, so the easiest fix is
to simply not refresh.

This fixes #19.