Change rules for populating Leads and Owners
When assigning leads and owners we generally don't want to fall through from
arch to os to area when a given arch or os doesn't have assigned leads or
owners.

However, for ownership we only want to consider arch and os if their labels
are actually mapped.

At the same time, we don't want to filter those out from the Architectures and
OperatingSystem properties because those control `os:xxx` and `arch:xxx`
terms being added to the trie which we still want even if those aren't
officially mapped.